+++ relay-server/src/actors/project.rs
@@ -21,7 +21,7 @@ use crate::actors::project_cache::{FetchProjectState, ProjectCache, ProjectError
 use crate::envelope::Envelope;
 use crate::extractors::RequestMeta;
 use crate::metrics::RelayCounters;
-use crate::utils::{ActorResponse, EnvelopeLimiter, Response};
+use crate::utils::{ActorResponse, EnvelopeLimiter, Response, SamplingConfig};
 
 /// The current status of a project state. Return value of `ProjectState::outdated`.
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
@@ -60,6 +60,9 @@ pub struct ProjectConfig {
     pub event_retention: Option<u16>,
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub quotas: Vec<Quota>,
+    /// Configuration for sampling traces, if not present there will be no sampling.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub sampling: Option<SamplingConfig>,
 }
 
 impl Default for ProjectConfig {
@@ -73,6 +76,7 @@ impl Default for ProjectConfig {
             datascrubbing_settings: DataScrubbingConfig::default(),
             event_retention: None,
             quotas: Vec::new(),
+            sampling: None,
         }
     }
 }
@@ -551,6 +555,29 @@ impl Actor for Project {
 
     fn stopped(&mut self, _ctx: &mut Self::Context) {
         log::debug!("project {} removed from cache", self.public_key);
+    }
+}
+
+/// Returns the project state if it is already cached.
+///
+/// This is used for cases when we only want to perform operations that do
+/// not require waiting for network requests.
+///
+pub struct GetCachedProjectState;
+
+impl Message for GetCachedProjectState {
+    type Result = Option<Arc<ProjectState>>;
+}
+
+impl Handler<GetCachedProjectState> for Project {
+    type Result = Option<Arc<ProjectState>>;
+
+    fn handle(
+        &mut self,
+        _message: GetCachedProjectState,
+        _context: &mut Context<Self>,
+    ) -> Self::Result {
+        self.state.clone()
     }
 }