+++ relay-server/src/endpoints/common.rs
@@ -12,12 +12,13 @@ use futures::prelude::*;
 use serde::Deserialize;
 
 use relay_common::{clone, metric, tryf, LogError};
+use relay_config::Config;
 use relay_general::protocol::{EventId, EventType};
 use relay_quotas::RateLimits;
 
 use crate::actors::events::{QueueEnvelope, QueueEnvelopeError};
 use crate::actors::outcome::{DiscardReason, Outcome, TrackOutcome};
-use crate::actors::project::CheckEnvelope;
+use crate::actors::project::{CheckEnvelope, Project};
 use crate::actors::project_cache::{GetProject, ProjectError};
 use crate::body::StorePayloadError;
 use crate::envelope::{AttachmentType, Envelope, EnvelopeError, ItemType, Items};
@@ -25,7 +26,6 @@ use crate::extractors::RequestMeta;
 use crate::metrics::RelayCounters;
 use crate::service::{ServiceApp, ServiceState};
 use crate::utils::{self, ApiErrorResponse, FormDataIter, MultipartError};
-use relay_config::Config;
 
 #[derive(Fail, Debug)]
 pub enum BadStoreRequest {
@@ -76,6 +76,9 @@ pub enum BadStoreRequest {
 
     #[fail(display = "event submission rejected with_reason: {:?}", _0)]
     EventRejected(DiscardReason),
+
+    #[fail(display = "envelope empty due to sampling")]
+    TraceSampled(Option<EventId>),
 }
 
 impl BadStoreRequest {
@@ -123,6 +126,9 @@ impl BadStoreRequest {
                     .longest_error()
                     .map(|r| Outcome::RateLimited(r.reason_code.clone()));
             }
+            //TODO fix this when we decide how to map empty Envelopes due to the trace being
+            //removed by sampling
+            BadStoreRequest::TraceSampled(_) => Outcome::Invalid(DiscardReason::Internal),
 
             // should actually never create an outcome
             BadStoreRequest::InvalidEventId => Outcome::Invalid(DiscardReason::Internal),
@@ -408,7 +414,7 @@ where
         .and_then(clone!(event_id, scoping, |project| {
             extract_envelope(&request, meta)
                 .into_future()
-                .and_then(clone!(project, |envelope| {
+                .and_then(clone!(project, event_id, |envelope| {
                     event_id.replace(envelope.event_id());
 
                     project
@@ -433,11 +439,62 @@ where
                         Err(BadStoreRequest::PayloadError(StorePayloadError::Overflow))
                     }
                 }))
-                .and_then(move |(envelope, rate_limits)| {
+                .and_then(clone!(event_manager, project_manager, |(
+                    envelope,
+                    rate_limits,
+                )| {
+                    type RetVal = ResponseFuture<
+                        (Envelope, RateLimits, Option<Addr<Project>>),
+                        BadStoreRequest,
+                    >;
+                    // do dynamic sampling on transactions
+                    let trace_context = match envelope.trace_context() {
+                        None => {
+                            // if we don't have a trace context we can't do dynamic sampling so stop.
+                            return Box::new(Ok((envelope, rate_limits, None)).into_future())
+                                as RetVal;
+                        }
+                        Some(trace_context) => trace_context,
+                    };
+                    // Sample and potentially remove transactions from the envelope.
+                    // We only do this if the envelope contains only transactions,
+                    // The reason for that is that in case of envelopes containing only
+                    // transactions we have a chance to end processing here (if the transactions
+                    // are sampled out).
+                    // If the envelope contains other items (beyond transactions then we cannot
+                    // shortcut the processing here so we'll do it after we queue the envelope).
+                    let response = project_manager
+                        .send(GetProject {
+                            public_key: trace_context.public_key,
+                        })
+                        // deal with mailbox errors
+                        .map_err(BadStoreRequest::ScheduleFailed)
+                        // do the fast path transaction sampling (if we can't do it here
+                        // we'll try again after the envelope is queued)
+                        .map(|project| (envelope, rate_limits, Some(project)));
+                    Box::new(response) as RetVal
+                }))
+                .and_then(|(envelope, rate_limits, sampling_project)| {
+                    // do the fast path transaction sampling (if we can't do it here
+                    // we'll try again after the envelope is queued)
+                    let event_id = envelope.event_id();
+
+                    utils::sample_transaction(envelope, sampling_project.clone(), true).then(
+                        move |result| match result {
+                            Err(()) => Err(BadStoreRequest::TraceSampled(event_id)),
+                            Ok(envelope) if envelope.is_empty() => {
+                                Err(BadStoreRequest::TraceSampled(event_id))
+                            }
+                            Ok(envelope) => Ok((envelope, rate_limits, sampling_project)),
+                        },
+                    )
+                })
+                .and_then(move |(envelope, rate_limits, sampling_project)| {
                     event_manager
                         .send(QueueEnvelope {
                             envelope,
                             project,
+                            sampling_project,
                             start_time,
                         })
                         .map_err(BadStoreRequest::ScheduleFailed)
@@ -469,6 +526,11 @@ where
 
             if !emit_rate_limit && matches!(error, BadStoreRequest::RateLimited(_)) {
                 return Ok(create_response(*event_id.borrow()));
+            }
+
+            if let BadStoreRequest::TraceSampled(event_id) = error {
+                log::debug!("creating response for trace sampled event");
+                return Ok(create_response(event_id));
             }
 
             let response = error.error_response();
@@ -542,7 +604,7 @@ mod tests {
             minimal,
             MinimalEvent {
                 id: None,
-                ty: EventType::Default
+                ty: EventType::Default,
             }
         );
     }
@@ -555,7 +617,7 @@ mod tests {
             minimal,
             MinimalEvent {
                 id: Some("037af9ac1b49494bacd7ec5114f801d9".parse().unwrap()),
-                ty: EventType::Default
+                ty: EventType::Default,
             }
         );
     }